[feat]: 관심사에 따른 컴포넌트 분리
