+++ README.md
@@ -74,6 +74,10 @@
 5. UI 구성 후 비지니스 로직도 하나씩 옮기기 시작.
 6. 전역변수는 useState로 변환.
 7. main에서 실행된 순서대로 App 컴포넌트에 옮겨 구현.
+8. UI와 비지니스 로직 모두 옮김
+9. UI들을 관심사에 따라 분리(id로 이미 나눠져 있어 편하게 분리함).
+10. 하지만 비지니스 로직에 데이터들이 서로 연결되어있는것들이 많아 복잡해 보임 -> 비지니스 로직들을 커스텀훅으로 빼보기로 함.
+11.
 
 ### 리팩토링 체크리스트
