+++ src/advanced/src/components/CartItem.tsx
@@ -1,0 +1,60 @@
+import React from 'react';
+import { Cart, Product } from '../App';
+
+const CartItem = ({ cart, productList, setCart, setProductList }) => {
+  const handleQuantityChange = (productId: string, change: number) => {
+    if ((cart[productId] || 0) + change < 0) {
+      return;
+    }
+
+    setCart((prevCart: Cart) => {
+      const newCart = { ...prevCart, [productId]: (prevCart[productId] || 0) + change };
+      if (newCart[productId] <= 0) {
+        delete newCart[productId];
+      }
+
+      return newCart;
+    });
+
+    setProductList((prevList: Product[]) =>
+      prevList.map((prev: Product) => (prev.id === productId ? { ...prev, quantity: prev.quantity - change } : prev)),
+    );
+  };
+
+  return (
+    <div id="cart-items">
+      {Object.entries(cart).map(([productId, quantity]) => {
+        const product = productList.find((item) => item.id === productId);
+        return (
+          <div key={productId} className="flex justify-between items-center mb-2">
+            <span>
+              {product?.name} - {product?.price}원 x {quantity}
+            </span>
+            <div>
+              <button
+                className="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1"
+                onClick={() => handleQuantityChange(productId, -1)}
+              >
+                -
+              </button>
+              <button
+                className="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1"
+                onClick={() => handleQuantityChange(productId, 1)}
+              >
+                +
+              </button>
+              <button
+                className="remove-item bg-red-500 text-white px-2 py-1 rounded"
+                onClick={() => handleQuantityChange(productId, -1 * quantity)}
+              >
+                삭제
+              </button>
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export default CartItem;
